requirements.txt
streamlit
openai
numpy
faiss-cpu
python-dotenv
langchain
pandas
streamlit
openai
python-dotenv
faiss-cpu
fastapi==0.110.0
uvicorn==0.27.0
numpy
langchain
langchain_openai
fastapi
uvicorn
faiss-cpu
numpy
openai
python-dotenv
redis
gunicorn
langchain_openai
